requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
openai>=1.3.0
requests>=2.31.0
duckduckgo-search>=3.9.0
PyMuPDF>=1.24.0
pypdf>=4.0.0
python-docx>=0.8.11
unstructured>=0.10.0
python-dotenv>=1.0.0
watchdog>=3.0.0
google-generativeai